+++ packages/camera/CHANGELOG.md
@@ -1,3 +1,21 @@
+## 0.2.9
+
+* Save photo orientation data on iOS.
+
+## 0.2.8
+
+* Add access to the image stream from Dart.
+* Use `cameraController.startImageStream(listener)` to process the images.
+
+## 0.2.7
+
+* Fix issue with crash when the physical device's orientation is unknown.
+
+## 0.2.6
+
+* Update the camera to use the physical device's orientation instead of the UI
+  orientation on Android.
+
 ## 0.2.5
 
 * Fix preview and video size with satisfying conditions of multiple outputs.