+++ packages/google_maps_flutter/CHANGELOG.md
@@ -1,3 +1,20 @@
+## 0.1.0
+
+* Move the map options from the GoogleMapOptions class to GoogleMap widget parameters.
+
+## 0.0.3+3
+
+* Relax Flutter version requirement to 0.11.9.
+
+## 0.0.3+2
+
+* Update README to recommend using the package from pub.
+
+## 0.0.3+1
+
+* Bug fix: custom marker images were not working on iOS as we were not keeping
+  a reference to the plugin registrar so couldn't fetch assets.
+
 ## 0.0.3
 
 * Don't export `dart:async`.